# Cherry-pick Reuse the epoll object to fix the dispatcher performance